Improve Perft
. use long rather than int to compute the number of moves in perft
. At depth 1 of performance test, we do not need to actually do and undo the moves,
  we only need to count them.

This drastically improves performance numbers for moves generation,
as we spend most of the time generating the moves, rather than moving the pieces.

+++ source/Perft.h
@@ -10,8 +10,8 @@
 #include "Common.h"
 #include "Board.h"
 
-int perft(class Board &board, int depth);
-int perft_fen(string fen, int depth);
+long perft(class Board &board, int depth);
+long perft_fen(string fen, int depth);
 void perft_benchmark(string fen, int depth);
 
 #endif /* PERFT_H */
